Extract dashboard events into events utils
